pyproject.toml
[tool.poetry]
name = "hw-drf"
version = "0.1.0"
description = ""
authors = ["alexander anciferov <sania.antsiferoff@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
python-dotenv = "^1.0.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
